\name{RGenPred-package}
\alias{RGenPred-package}
\alias{RGenPred}
\docType{package}
\title{
\packageTitle{RGenPred}
}
\description{
\packageDescription{RGenPred}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{RGenPred}

Maintainer: \packageMaintainer{RGenPred}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RGenPred)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RGenPred}}
%\section{Documentation Index}{\packageIndices{RGenPred}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
